+++ src/main/kotlin/insight/generation/EventListenerGenerationSupport.kt
@@ -18,18 +18,21 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.demonwav.mcdev.platform.velocity.util
+package com.demonwav.mcdev.insight.generation
 
-import com.demonwav.mcdev.util.SemanticVersion
+import com.intellij.openapi.editor.Editor
+import com.intellij.psi.PsiClass
+import com.intellij.psi.PsiElement
 
-object VelocityConstants {
+interface EventListenerGenerationSupport {
 
-    const val PLUGIN_ANNOTATION = "com.velocitypowered.api.plugin.Plugin"
-    const val SUBSCRIBE_ANNOTATION = "com.velocitypowered.api.event.Subscribe"
-    const val POST_ORDER = "com.velocitypowered.api.event.PostOrder"
-    const val KYORI_TEXT_COLOR = "net.kyori.text.format.TextColor"
+    fun canGenerate(context: PsiElement, editor: Editor): Boolean
 
-    val API_2 = SemanticVersion.release(2)
-    val API_3 = SemanticVersion.release(3)
-    val API_4 = SemanticVersion.release(4)
+    fun generateEventListener(
+        context: PsiElement,
+        listenerName: String,
+        eventClass: PsiClass,
+        data: GenerationData?,
+        editor: Editor
+    )
 }