+++ src/main/kotlin/insight/generation/ui/EventGenerationPanel.kt
@@ -43,12 +43,10 @@ open class EventGenerationPanel(val chosenClass: PsiClass) {
     /**
      * This is called when the dialog is closing from the OK action. The platform should fill in their [GenerationData] object as
      * needed for whatever information their panel provides. The state of the panel can be assumed to be valid, since this will only be
-     * called if [.doValidate] has passed successfully.
+     * called if [doValidate] has passed successfully.
 
      * @return The [GenerationData] object which will be passed to the
-     * * [AbstractModule#doPreEventGenerate()][com.demonwav.mcdev.platform.AbstractModule.doPreEventGenerate] and
-     * * [AbstractModule#generateEventListenerMethod][com.demonwav.mcdev.platform.AbstractModule.generateEventListenerMethod]
-     * * methods.
+     * [EventListenerGenerationSupport][com.demonwav.mcdev.insight.generation.EventListenerGenerationSupport]
      */
     open fun gatherData(): GenerationData? {
         return null