+++ src/main/kotlin/platform/velocity/VelocityModule.kt
@@ -22,16 +22,12 @@ package com.demonwav.mcdev.platform.velocity
 
 import com.demonwav.mcdev.asset.PlatformAssets
 import com.demonwav.mcdev.facet.MinecraftFacet
-import com.demonwav.mcdev.insight.generation.GenerationData
+import com.demonwav.mcdev.insight.generation.EventListenerGenerationSupport
 import com.demonwav.mcdev.platform.AbstractModule
 import com.demonwav.mcdev.platform.PlatformType
-import com.demonwav.mcdev.platform.velocity.generation.VelocityGenerationData
 import com.demonwav.mcdev.platform.velocity.util.VelocityConstants
-import com.demonwav.mcdev.platform.velocity.util.VelocityConstants.SUBSCRIBE_ANNOTATION
-import com.demonwav.mcdev.util.createVoidMethodWithParameterType
 import com.demonwav.mcdev.util.runCatchingKtIdeaExceptions
 import com.intellij.lang.jvm.JvmModifier
-import com.intellij.psi.JavaPsiFacade
 import com.intellij.psi.PsiClass
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiMethod
@@ -44,33 +40,9 @@ class VelocityModule(facet: MinecraftFacet) : AbstractModule(facet) {
     override val type = PlatformType.VELOCITY
     override val icon = PlatformAssets.VELOCITY_ICON
 
+    override val eventListenerGenSupport: EventListenerGenerationSupport = VelocityEventListenerGenerationSupport()
+
     override fun isEventClassValid(eventClass: PsiClass, method: PsiMethod?): Boolean = true
-
-    override fun generateEventListenerMethod(
-        containingClass: PsiClass,
-        chosenClass: PsiClass,
-        chosenName: String,
-        data: GenerationData?,
-    ): PsiMethod? {
-        val method = createVoidMethodWithParameterType(project, chosenName, chosenClass) ?: return null
-        val modifierList = method.modifierList
-
-        val subscribeAnnotation = modifierList.addAnnotation(SUBSCRIBE_ANNOTATION)
-
-        val generationData = data as VelocityGenerationData
-
-        if (generationData.eventOrder != "NORMAL") {
-            val value = JavaPsiFacade.getElementFactory(project)
-                .createExpressionFromText(
-                    "com.velocitypowered.api.event.PostOrder." + generationData.eventOrder,
-                    subscribeAnnotation,
-                )
-
-            subscribeAnnotation.setDeclaredAttributeValue("order", value)
-        }
-
-        return method
-    }
 
     override fun shouldShowPluginIcon(element: PsiElement?): Boolean {
         val identifier = element?.toUElementOfType<UIdentifier>()